feat(product): #5063 库存米数小数化（1 位小数）—— 三层 INTEGER 一并升级 (#5147)
* feat(product): #5063 库存米数小数化（1 位小数）—— 三层 INTEGER 一并升级

* fix(cases): #5063 用例 data_checks 里的裸双引号未转义（PyYAML 严格解析失败 ⇒ drift_audit 的 mutable-locator 判据整条报错、9 条 component|* 豁免被误判「已归零」）

* chore(cases): #5063 重锚 case-machine 读数（53→60）—— 新增 7 条 [backend-contract] 用例，真债务 no_channel_total 仍为 0

---------

+++ backend/admin-api/src/main/java/com/migao/admin/dto/ProductResponse.java
@@ -74,7 +74,7 @@ public class ProductResponse {
     /**
      * 库存数量
      */
-    private Integer stock;
+    private BigDecimal stock;
 
     /**
      * 库存预警阈值
@@ -109,12 +109,12 @@ public class ProductResponse {
     /**
      * SKU总库存
      */
-    private Integer totalStock;
+    private BigDecimal totalStock;
 
     /**
      * 累计销量
      */
-    private Integer salesCount;
+    private BigDecimal salesCount;
 
     /**
      * 累计销售额
